Added path validation in Cloudinary.js

+++ backend/src/utils/cloudinary.js
@@ -10,8 +10,11 @@ cloudinary.config({
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
+        if (!localFilePath) return null;
+
         const response = await cloudinary.uploader.upload(localFilePath);
         fs.unlinkSync(localFilePath);
+
         return response.secure_url;
     } catch (error) {
         console.log(error.message);
